[gradebook-service] Changed currentYear data type in SemesterService

+++ services/gradebook-service/src/main/java/com/rebirthlab/gradebook/application/service/semester/SemesterServiceImpl.java
@@ -62,7 +62,7 @@ public class SemesterServiceImpl implements SemesterService {
 
     @Override
     public List<Semester> findActual() {
-        int currentYear = LocalDate.now().getYear();
+        short currentYear = (short) LocalDate.now().getYear();
         return semesterRepository.findActual(currentYear);
     }
